+++ tools/replay/route.cc
@@ -4,7 +4,7 @@
 #include <filesystem>
 #include <regex>
 
-#include "third_party/json11/json11.hpp"
+#include "json11/json11.hpp"
 #include "system/hardware/hw.h"
 #include "tools/replay/py_downloader.h"
 #include "tools/replay/replay.h"